requirements.txt
langchain==1.2.15
langchain-core==1.2.28
langchain-openrouter==0.2.1
pydantic==2.12.5
python-dotenv==1.2.2
